+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * Go: Add `LCS` ([#3475](https://github.com/valkey-io/valkey-glide/pull/3475))
 * Core/FFI/Go: Add Support for Async and Sync Client Types in FFI ([#3451](https://github.com/valkey-io/valkey-glide/pull/3451))
 * Core/FFI/Go: Add support for async and sync client types in FFI ([#3451](https://github.com/valkey-io/valkey-glide/pull/3451))
+* Go: Add `Config Rewrite` ([#3156](https://github.com/valkey-io/valkey-glide/pull/3156))
 
 #### Breaking Changes